+++ plugins/kafka/src/Router.tsx
@@ -18,7 +18,6 @@ import { Entity } from '@backstage/catalog-model';
 import React from 'react';
 import { Route, Routes } from 'react-router';
 import { useEntity } from '@backstage/plugin-catalog-react';
-import { rootCatalogKafkaRouteRef } from './plugin';
 import { KAFKA_CONSUMER_GROUP_ANNOTATION } from './constants';
 import { KafkaTopicsForConsumer } from './components/ConsumerGroupOffsets/ConsumerGroupOffsets';
 import { MissingAnnotationEmptyState } from '@backstage/core-components';
@@ -44,10 +43,7 @@ export const Router = (_props: Props) => {
 
   return (
     <Routes>
-      <Route
-        path={`${rootCatalogKafkaRouteRef.path}`}
-        element={<KafkaTopicsForConsumer />}
-      />
+      <Route path="/" element={<KafkaTopicsForConsumer />} />
     </Routes>
   );
 };